# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JamJam Trek API",
        "description": "API documentation for JamJam Trek application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://161.97.167.73:8001",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/reviews/publishable": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get approved/publishable reviews",
                "description": "Retrieves only approved reviews (status=1) with pagination, ordered by most recent",
                "operationId": "getPublishableReviews",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of reviews per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 8,
                            "example": 8
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publishable reviews fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Publishable reviews fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    },
                                                    "review": {
                                                        "type": "string",
                                                        "example": "Great service!"
                                                    },
                                                    "rating": {
                                                        "type": "number",
                                                        "example": 4.5
                                                    },
                                                    "trek": {
                                                        "type": "string",
                                                        "example": "Everest Base Camp Trek",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-01-30T12:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 24
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No publishable reviews found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/reviews/{id}": {
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "description": "Permanently deletes a review by ID",
                "operationId": "deleteReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reviews deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/reviews/stats": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get review statistics",
                "description": "Returns count of positive reviews (rating >= 4) and negative reviews (rating < 3)",
                "operationId": "getPositiveAndNegativeReviewsCount",
                "responses": {
                    "200": {
                        "description": "Review statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "positive_reviews": {
                                                    "description": "Count of reviews with rating >= 4",
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "negative_reviews": {
                                                    "description": "Count of reviews with rating < 3",
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}